+++ Runtime/Leaderboards/LocalizedTitles.cs
@@ -1,0 +1,32 @@
+using System;
+
+namespace YandexGames.Leaderboards
+{
+    /// <summary>
+    /// Represents localized leaderboard titles
+    /// </summary>
+    [Serializable]
+    public class LocalizedTitles
+    {
+        /// <summary>English title</summary>
+        public string en;
+        
+        /// <summary>Russian title</summary>
+        public string ru;
+        
+        /// <summary>Turkish title</summary>
+        public string tr;
+        
+        /// <summary>German title</summary>
+        public string de;
+        
+        /// <summary>French title</summary>
+        public string fr;
+        
+        /// <summary>Spanish title</summary>
+        public string es;
+        
+        /// <summary>Portuguese title</summary>
+        public string pt;
+    }
+}