+++ Runtime/YandexGamesCallbackReceiver.cs
@@ -74,5 +74,57 @@ namespace YandexGames
         {
             YandexGames.OnRewardedAdError(error);
         }
+
+        // Leaderboards callbacks
+        public void OnSetLeaderboardScoreComplete()
+        {
+            YandexGames.OnSetLeaderboardScoreComplete();
+        }
+
+        public void OnSetLeaderboardScoreError(string error)
+        {
+            YandexGames.OnSetLeaderboardScoreError(error);
+        }
+
+        public void OnGetLeaderboardDescriptionComplete(string json)
+        {
+            YandexGames.OnGetLeaderboardDescriptionComplete(json);
+        }
+
+        public void OnGetLeaderboardDescriptionError(string error)
+        {
+            YandexGames.OnGetLeaderboardDescriptionError(error);
+        }
+
+        public void OnGetLeaderboardPlayerEntryComplete(string json)
+        {
+            YandexGames.OnGetLeaderboardPlayerEntryComplete(json);
+        }
+
+        public void OnGetLeaderboardPlayerEntryError(string error)
+        {
+            YandexGames.OnGetLeaderboardPlayerEntryError(error);
+        }
+
+        public void OnGetLeaderboardEntriesComplete(string json)
+        {
+            YandexGames.OnGetLeaderboardEntriesComplete(json);
+        }
+
+        public void OnGetLeaderboardEntriesError(string error)
+        {
+            YandexGames.OnGetLeaderboardEntriesError(error);
+        }
+
+        // Remote Config callbacks
+        public void OnGetFlagsComplete(string json)
+        {
+            YandexGames.OnGetFlagsComplete(json);
+        }
+
+        public void OnGetFlagsError(string error)
+        {
+            YandexGames.OnGetFlagsError(error);
+        }
     }
 }